// src/modules/disposal-points/enums/dp-entity-properties-names.enum.ts
/* DisposalPointEntityPropertiesNamesEnum */
export enum DisposalPointEntityPropertiesNamesEnum {
  ACCEPTED_WASTE_CATEGORY = "dpAcceptedWasteCategory",
  ID = "dpId",
  LATITUDE = "dpLatitude",
  LOCALITY_NAME = "dpLocalityName",
  LOCALITY_TYPE = "dpLocalityType",
  LONGITUDE = "dpLongitude",
  NEIGHBORHOOD = "dpNeighborhood",
}